feat(TDKN-126): Remove title on JSONSchema Object. (#141)
* feat(TDKN-126): Remove title on JSONSchema Object.

This commit will remove title on JSONSchema object when the object is not
present on the current form.
This will allow to not display empty object.

* feat(TDKN-126): Change the title value to an internationalized value.

The getTitle is not a valid value, since this is not supposed to be used
as the title of the component. It was changed to getDisplayName.